requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==5.1.6
PyMySQL==1.1.1
django-cors-headers==4.9.0
django-jazzmin==3.0.4
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
gunicorn==25.3.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.10
requests==2.33.1
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
sentry-sdk==2.1.1
django-resized==1.0.2
django-pwa==1.1.0
djangorestframework-simplejwt==5.3.1
